backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.34
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0